==> scripts/03 Data_Visualisation.Rmd <==
---
title: "Data_Visualisation"
author: "Leonie Rettig & Filip Gonschorek & Steffen Zeeb"
output: html_document
---

```{r setup, include=F, message=F, results=F}
rm(list=ls())
# Load and potentially install all the packages for this lab
#p_needed <- c("tidyverse", "dplyr", "ggplot2")
#packages <- rownames(installed.packages())
#p_to_install <- p_needed[!(p_needed %in% packages)]
#if (length(p_to_install) > 0) {
#  install.packages(p_to_install)
#}
#lapply(p_needed, require, character.only = TRUE)
```

```{r}
load("..processed-data/overview.RData")
load("..processed-data/Landkreis_liste.RData")
#load("C:/Users/Leoni/OneDrive - bwedu/GitHub/WirVsVirus_1_043_C_Medizinische_Versorgung_ForecastFallzahlen/processed-data/overview.RData")
#load("C:/Users/Leoni/OneDrive - bwedu/GitHub/WirVsVirus_1_043_C_Medizinische_Versorgung_ForecastFallzahlen/processed-data/Landkreis_liste.RData")
```
## Setup for Data Visualisation
```{r Set default theme, results=F, message=F}
mytheme <- 
  theme_minimal() +
  theme(axis.line = element_line(colour = "ivory4"),
        axis.text.x = element_text(size = 12,
                                   hjust = .5),
        axis.ticks = element_line(colour = "ivory4"),
        axis.text.y = element_text(size = 12),
        axis.title = element_text(size = 12),
        axis.title.x = element_blank(),
        plot.caption = element_text(size = 12),
        panel.background = element_blank(),
        panel.grid.major.y = element_line(colour = adjustcolor("ivory4", alpha.f = .35),
                                          linetype = "dashed"),
        panel.grid.minor.y = element_blank(),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        legend.position = "none",
        plot.title = element_text(size = 12,
                                  hjust = 0,
                                  face = "bold")
  )

theme_set(mytheme)
```

```{r Base Plot}
for ( i in 1:20){
k <- Landkreis_liste[[overview$num[i]]]

plot(1:dim(k)[1], k$Fallgesamt, type = "l", col = "blue", ylim = c(0,max(c(k$predicted_values,k$Fallgesamt))+5), ylab = paste("Faelle",names(Landkreis_liste)[i], sep = " "), xlab = "Tage seit Erstinfektion")
lines(1:dim(k)[1], k$predicted_values, lty = 2 )
name <- paste0(i,".png")
png(filename = paste0("../figures/",name),)
#png(filename = paste0("C:/Users/Leoni/OneDrive - bwedu/GitHub/WirVsVirus_1_043_C_Medizinische_Versorgung_ForecastFallzahlen/figures/", name))
}
```

```{r GGPlot, warning=F}
for ( i in 1:20){
k <- Landkreis_liste[[overview$num[i]]]
x <- 1:dim(k)[1]
y <- k$Fallgesamt

plot <- ggplot(NULL,aes(x,y)) +
  geom_line(aes(color = "Observed Values"), linetype=2) +
  ylab(paste("Faelle",names(Landkreis_liste)[overview$num[i]], sep = " ")) + 
  xlab("Tage seit Erstinfektion") +
  scale_y_continuous(
    limits = c(0,max(c(k$predicted_values,k$Fallgesamt))+5)
  )+
  geom_line(aes(1:dim(k)[1], k$predicted_values, color="Predicted Values"))+
theme(
    axis.title.x = element_text(size = 12),
    legend.position = c(0.83, 0.85),
    legend.key.size = unit(0.3, "cm"),
    legend.text = element_text(size = 12),
    legend.title = element_blank()
)

name <- paste0(20+i,".png")
png(filename =paste0("../figures/ggplot/",name))
#png(filename =paste0("C:/Users/Leoni/OneDrive - bwedu/GitHub/WirVsVirus_1_043_C_Medizinische_Versorgung_ForecastFallzahlen/figures/",name))
plot
dev.off()
}
```
